# backend/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "fratelli-feccia backend API",
        "title": "fratelli-feccia API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/vehicles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "List vehicles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by targa",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.VehicleResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Create vehicle",
                "parameters": [
                    {
                        "description": "Vehicle data",
                        "name": "vehicle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VehicleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.VehicleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/vehicles/gps-position-by-plate/{targa}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update vehicle GPS position by targa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle targa",
                        "name": "targa",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "GPS position",
                        "name": "position",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VehicleGPSUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GPSUpdateResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/vehicles/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update vehicle (full replace of the create-able fields)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vehicle data",
                        "name": "vehicle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VehicleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.VehicleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Delete vehicle (logical, sets active=false)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/vehicles/{id}/gps-history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "GPS history for a vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID (UUID) or targa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GPSHistoryResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/vehicles/{id}/gps-position": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update vehicle GPS position (by id or targa)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID (UUID) or targa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "GPS position",
                        "name": "position",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VehicleGPSUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GPSUpdateResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/vehicles/{id}/temperature": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Temperature history for a vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only out-of-range readings",
                        "name": "only_alerts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.TemperatureReadingResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/vehicles/{id}/temperature-thresholds": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Set temperature thresholds for a vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thresholds",
                        "name": "thresholds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TemperatureThresholdsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TemperatureThresholdsResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/webhooks/gps/{vendor}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "GPS provider webhook ingestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GPS vendor",
                        "name": "vendor",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Normalized GPS payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GPSWebhookPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GPSUpdateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/webhooks/temperature/{vendor}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Temperature sensor webhook ingestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sensor vendor",
                        "name": "vendor",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Temperature reading",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TemperatureWebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TemperatureWebhookResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.GPSHistoryResponse": {
            "type": "object",
            "properties": {
                "gps_source": {
                    "type": "string"
                },
                "heading": {
                    "type": "number"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "speed_kmh": {
                    "type": "number"
                },
                "targa": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "vehicle_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.GPSPositionShort": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
        "dto.GPSUpdateResult": {
            "type": "object",
            "properties": {
                "gps_source": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "position": {
                    "$ref": "#/definitions/dto.GPSPositionShort"
                },
                "targa": {
                    "type": "string"
                }
            }
        },
...
        "dto.TemperatureReadingResponse": {
            "type": "object",
            "properties": {
                "out_of_range": {
                    "type": "boolean"
                },
                "sensor_id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "targa": {
                    "type": "string"
                },
                "temp_celsius": {
                    "type": "number"
                },
                "ts": {
                    "type": "string"
                },
                "vehicle_id": {
                    "type": "string"
                }
            }
        },
        "dto.TemperatureThresholdsRequest": {
            "type": "object",
            "properties": {
                "temp_max": {
                    "type": "number"
                },
                "temp_min": {
                    "type": "number"
                }
            }
        },
...
        "dto.VehicleGPSUpdateRequest": {
            "type": "object",
            "required": [
                "lat",
                "lng"
            ],
            "properties": {
                "heading": {
                    "type": "number"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "speed_kmh": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "dto.VehicleRequest": {
            "type": "object",
            "required": [
                "targa"
            ],
            "properties": {
                "anno": {
                    "type": "integer"
                },
                "gps_api_key": {
                    "type": "string"
                },
                "gps_tracker_tipo": {
                    "type": "string"
                },
                "gps_tracker_url": {
                    "type": "string"
                },
                "marca": {
                    "type": "string"
                },
                "modello": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "portata_kg": {
                    "type": "number"
                },
                "scompartature": {
                    "type": "integer"
                },
                "targa": {
                    "type": "string"
                },
                "tipo_veicolo": {
                    "type": "string"
                }
            }
        },
        "dto.VehicleResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "anno": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "gps_active": {
                    "type": "boolean"
                },
                "gps_api_key": {
                    "type": "string"
                },
                "gps_source": {
                    "type": "string"
                },
                "gps_tracker_tipo": {
                    "type": "string"
                },
                "gps_tracker_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_gps_update": {
                    "type": "string"
                },
                "last_heading": {
                    "type": "number"
                },
                "last_lat": {
                    "type": "number"
                },
                "last_lng": {
                    "type": "number"
                },
                "last_speed_kmh": {
                    "type": "number"
                },
                "last_temp_alert": {
                    "type": "boolean"
                },
                "last_temp_celsius": {
                    "type": "number"
                },
                "last_temp_sensor_id": {
                    "type": "string"
                },
                "last_temp_ts": {
                    "type": "string"
                },
                "marca": {
                    "type": "string"
                },
                "modello": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "portata_kg": {
                    "type": "number"
                },
                "scompartature": {
                    "type": "integer"
                },
                "targa": {
                    "type": "string"
                },
                "temp_max": {
                    "type": "number"
                },
                "temp_min": {
                    "type": "number"
                },
                "tipo_veicolo": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}